api/requirements.txt
azure-functions
azure-core==1.32.0
azure-identity==1.21.0
azure-storage-blob==12.25.1

numpy==1.26.4
matplotlib==3.8.3
tensorflow==2.16.2
tf_keras==2.16.0
requests==2.32.3